requirements.txt
pandas
googletrans
tqdm
GitPython